Refactor grouping logic and add Grouping.cs
Moved grouping methods from Rejigs.cs to a new Grouping.cs file for better separation of concerns. Renamed test file and class to GroupingTests for clarity.
