+++ Rejigs.Tests/GroupingTests.cs
@@ -2,15 +2,15 @@ using NUnit.Framework;
 
 namespace Rejigs.Tests;
 
-public class Grouping
+public class GroupingTests
 {
     [Test]
     public void Grouping_WithQuantifier()
     {
         var regex = Rejigs.Create()
-            .Grouping(r => r.Text("abc"))
-            .Exactly(2)
-            .Build();
+                          .Grouping(r => r.Text("abc"))
+                          .Exactly(2)
+                          .Build();
 
         Assert.That("abcabc", Does.Match(regex));
         Assert.That("abc", Does.Not.Match(regex));
